PTHREAD_STACK_MIN: bump to a real multiple of every target page size
Real, unmodified musl upstream defines PTHREAD_STACK_MIN as 2048 -- not
a multiple of any real page size (4 KiB here; 16 KiB/64 KiB on future
ports this project's multi-arch goal targets). The Open POSIX Test
Suite's own threads_scenarii.c helper (shared by pthread_create/
pthread_detach/pthread_exit/pthread_join) bails out UNTESTED whenever
sysconf(_SC_THREAD_STACK_MIN) isn't page-aligned -- found by comparing
a full-corpus pilot run against the same corpus on a real glibc host,
which never trips this (its own PTHREAD_STACK_MIN, 16384, is aligned).
19 real files were bailing out before testing anything at all.

Bumped to 65536, a multiple of every page size this project has ever
cared about, so the fix holds across future architecture ports too,
not just this one.

That alone measured zero effect: sysconf.c's own values[] table stored
every _SC_* entry as a 16-bit short, silently truncating 65536 to 0.
Widened to int -- confirmed by inspection that every other entry (the
JT/RLIM sentinel-encoded ones included) already fit comfortably within
int, so this doesn't change any other sysconf() return value.

Verified: 15 of 19 previously-UNTESTED files now genuinely PASS
against this kernel; real pthread_create/join still works. Confirmed
via an isolated A/B run that the pre-existing pthread_cond_broadcast/
{1-2,2-3,4-2}.c TIMEOUTs (a 10,000-thread stress test) are unaffected
by this value either way.

+++ src/conf/sysconf.c
@@ -27,7 +27,21 @@
 
 long sysconf(int name)
 {
-	static const short values[] = {
+	/* OxideBSD patch: widened `short` -> `int`. Real, unmodified musl upstream never needed more
+	 * than 16 bits here -- every entry it ever stored fit comfortably. This project's own
+	 * `PTHREAD_STACK_MIN` (see include/limits.h's own doc comment: bumped to 65536 so it stays a
+	 * real multiple of every page size this project's multi-arch goal targets, 4 KiB/16 KiB/64
+	 * KiB, not just this port's own 4 KiB) doesn't fit in a signed 16-bit slot at all (max
+	 * positive value 32767) -- silently truncated to `(short)65536` = `0` before this widening,
+	 * confirmed live: `sysconf(_SC_THREAD_STACK_MIN)` returned `0` instead of `65536`, found by
+	 * comparing this kernel's own POSIX conformance pilot run against a real host run before and
+	 * after the `include/limits.h` change showed zero effect. Every other entry in this table
+	 * (the `JT`/`RLIM` sentinel-encoded ones included) already fits well within `int`'s own range,
+	 * so this widening changes nothing else's runtime behavior -- confirmed by inspecting every
+	 * macro below: `JT(x)`'s and `RLIM(x)`'s own bitwise-OR results are computed in `int`
+	 * arithmetic regardless of the target array's element width, and every other real value
+	 * assigned below already sat comfortably inside a 16-bit `short`. */
+	static const int values[] = {
 		[_SC_ARG_MAX] = JT_ARG_MAX,
 		[_SC_CHILD_MAX] = RLIM(NPROC),
 		[_SC_CLK_TCK] = 100,
